Rename narrowReducers to narrowsReducers.
